+++ src/main/java/bot/DialogBox.java
@@ -1,4 +1,4 @@
-package duke;
+package bot;
 
 import java.io.IOException;
 import java.util.Collections;
