+++ src/main/java/bot/TaskList.java
@@ -1,4 +1,4 @@
-package duke;
+package bot;
 
 import java.util.ArrayList;
 import java.util.List;
